removed router since its app based & added auto checker if ollama is running
